test(amplify-e2e-tests): split long running tests
