+++ packages/amplify-e2e-tests/src/__tests__/migration/api.key.migration1.test.ts
@@ -67,63 +67,4 @@ describe('amplify add api', () => {
       amplifyPushUpdate(projRoot, /Attempting to edit the local secondary index SomeLSI on the TodoTable table in the Todo stack.*/),
     ).rejects.toThrowError('Process exited with non zero exit code 1');
   });
-
-  it('init project, run invalid migration trying to add more than one gsi, and check for error', async () => {
-    const projectName = 'migratingkey';
-    const initialSchema = 'migrations_key/initial_schema.graphql';
-    const nextSchema1 = 'migrations_key/cant_add_more_gsi.graphql';
-    await initJSProjectWithProfile(projRoot, { name: projectName });
-    await addApiWithSchema(projRoot, initialSchema);
-    await amplifyPush(projRoot);
-    updateApiSchema(projRoot, projectName, nextSchema1);
-    await expect(
-      amplifyPushUpdate(
-        projRoot,
-        /Attempting to add more than 1 global secondary index SomeGSI1 and someGSI2 on the TodoTable table in the Todo stack.*/,
-      ),
-    ).rejects.toThrowError('Process exited with non zero exit code 1');
-  });
-
-  it('init project, run invalid migration trying to delete more than one gsi, and check for error', async () => {
-    const projectName = 'migratingkey1';
-    const initialSchema = 'migrations_key/initial_schema1.graphql';
-    const nextSchema1 = 'migrations_key/cant_remove_more_gsi.graphql';
-    await initJSProjectWithProfile(projRoot, { name: projectName });
-    await addApiWithSchema(projRoot, initialSchema);
-    await amplifyPush(projRoot);
-    updateApiSchema(projRoot, projectName, nextSchema1);
-    await expect(
-      amplifyPushUpdate(
-        projRoot,
-        /Attempting to delete more than 1 global secondary index SomeGSI1 and someGSI2 on the TodoTable table in the Todo stack.*/,
-      ),
-    ).rejects.toThrowError('Process exited with non zero exit code 1');
-  });
-
-  it('init project, run invalid migration trying to add and delete gsi, and check for error', async () => {
-    const projectName = 'migratingkey2';
-    const initialSchema = 'migrations_key/initial_schema.graphql';
-    const nextSchema1 = 'migrations_key/cant_update_delete_gsi.graphql';
-    await initJSProjectWithProfile(projRoot, { name: projectName });
-    await addApiWithSchema(projRoot, initialSchema);
-    await amplifyPush(projRoot);
-    updateApiSchema(projRoot, projectName, nextSchema1);
-    await expect(
-      amplifyPushUpdate(
-        projRoot,
-        /Attempting to add and delete a global secondary index SomeGSI1 and someGSI2 on the TodoTable table in the Todo stack.*/,
-      ),
-    ).rejects.toThrowError('Process exited with non zero exit code 1');
-  });
-
-  it('init project, run valid migration adding a GSI', async () => {
-    const projectName = 'validaddinggsi';
-    const initialSchema = 'migrations_key/initial_schema.graphql';
-    const nextSchema1 = 'migrations_key/add_gsi.graphql';
-    await initJSProjectWithProfile(projRoot, { name: projectName });
-    await addApiWithSchema(projRoot, initialSchema);
-    await amplifyPush(projRoot);
-    updateApiSchema(projRoot, projectName, nextSchema1);
-    await amplifyPushUpdate(projRoot, /GraphQL endpoint:.*/);
-  });
 });
