+++ packages/amplify-e2e-tests/src/__tests__/migration/api.key.migration2.test.ts
@@ -1,0 +1,74 @@
+import { initJSProjectWithProfile, deleteProject, amplifyPush, amplifyPushUpdate } from 'amplify-e2e-core';
+import { addApiWithSchema, updateApiSchema } from 'amplify-e2e-core';
+import { createNewProjectDir, deleteProjectDir } from 'amplify-e2e-core';
+
+describe('amplify add api', () => {
+  let projRoot: string;
+  beforeEach(async () => {
+    projRoot = await createNewProjectDir('api-key-migration');
+  });
+
+  afterEach(async () => {
+    await deleteProject(projRoot);
+    deleteProjectDir(projRoot);
+  });
+
+  it('init project, run invalid migration trying to add more than one gsi, and check for error', async () => {
+    const projectName = 'migratingkey';
+    const initialSchema = 'migrations_key/initial_schema.graphql';
+    const nextSchema1 = 'migrations_key/cant_add_more_gsi.graphql';
+    await initJSProjectWithProfile(projRoot, { name: projectName });
+    await addApiWithSchema(projRoot, initialSchema);
+    await amplifyPush(projRoot);
+    updateApiSchema(projRoot, projectName, nextSchema1);
+    await expect(
+      amplifyPushUpdate(
+        projRoot,
+        /Attempting to add more than 1 global secondary index SomeGSI1 and someGSI2 on the TodoTable table in the Todo stack.*/,
+      ),
+    ).rejects.toThrowError('Process exited with non zero exit code 1');
+  });
+
+  it('init project, run invalid migration trying to delete more than one gsi, and check for error', async () => {
+    const projectName = 'migratingkey1';
+    const initialSchema = 'migrations_key/initial_schema1.graphql';
+    const nextSchema1 = 'migrations_key/cant_remove_more_gsi.graphql';
+    await initJSProjectWithProfile(projRoot, { name: projectName });
+    await addApiWithSchema(projRoot, initialSchema);
+    await amplifyPush(projRoot);
+    updateApiSchema(projRoot, projectName, nextSchema1);
+    await expect(
+      amplifyPushUpdate(
+        projRoot,
+        /Attempting to delete more than 1 global secondary index SomeGSI1 and someGSI2 on the TodoTable table in the Todo stack.*/,
+      ),
+    ).rejects.toThrowError('Process exited with non zero exit code 1');
+  });
+
+  it('init project, run invalid migration trying to add and delete gsi, and check for error', async () => {
+    const projectName = 'migratingkey2';
+    const initialSchema = 'migrations_key/initial_schema.graphql';
+    const nextSchema1 = 'migrations_key/cant_update_delete_gsi.graphql';
+    await initJSProjectWithProfile(projRoot, { name: projectName });
+    await addApiWithSchema(projRoot, initialSchema);
+    await amplifyPush(projRoot);
+    updateApiSchema(projRoot, projectName, nextSchema1);
+    await expect(
+      amplifyPushUpdate(
+        projRoot,
+        /Attempting to add and delete a global secondary index SomeGSI1 and someGSI2 on the TodoTable table in the Todo stack.*/,
+      ),
+    ).rejects.toThrowError('Process exited with non zero exit code 1');
+  });
+
+  it('init project, run valid migration adding a GSI', async () => {
+    const projectName = 'validaddinggsi';
+    const initialSchema = 'migrations_key/initial_schema.graphql';
+    const nextSchema1 = 'migrations_key/add_gsi.graphql';
+    await initJSProjectWithProfile(projRoot, { name: projectName });
+    await addApiWithSchema(projRoot, initialSchema);
+    await amplifyPush(projRoot);
+    updateApiSchema(projRoot, projectName, nextSchema1);
+    await amplifyPushUpdate(projRoot, /GraphQL endpoint:.*/);
+  });
+});
